third_party/WebKit/LayoutTests/http/tests/serviceworker/fetch-event.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<body>
<script>
var worker = 'resources/fetch-event-test-worker.js';

async_test(function(t) {
    var scope = 'resources/simple.html?string';
    service_worker_unregister_and_register(t, worker, scope)
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(t.step_func(function(sw) {
          return wait_for_state(t, sw, 'activated');
        }))
      .then(t.step_func(function() { return with_iframe(scope); }))
      .then(t.step_func(function(frame) {
          assert_equals(
            frame.contentDocument.body.textContent,
            'Test string',
            'Service Worker should respond to fetch with a test string');
          assert_equals(
            frame.contentDocument.contentType,
            'text/plain',
            'The content type of the response created with a string should be text/plain');
          assert_equals(
            frame.contentDocument.characterSet,
            'UTF-8',
            'The character set of the response created with a string should be UTF-8');
          return service_worker_unregister_and_done(t, scope);
        }))
      .catch(unreached_rejection(t));
  }, 'Service Worker responds to fetch event with string');

async_test(function(t) {
    var scope = 'resources/simple.html?blob';
    service_worker_unregister_and_register(t, worker, scope)
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(t.step_func(function(sw) {
          return wait_for_state(t, sw, 'activated');
        }))
      .then(t.step_func(function() { return with_iframe(scope); }))
      .then(t.step_func(function(frame) {
          assert_equals(
            frame.contentDocument.body.textContent,
            'Test blob',
            'Service Worker should respond to fetch with a test string');
          return service_worker_unregister_and_done(t, scope);
        }))
      .catch(unreached_rejection(t));
  }, 'Service Worker responds to fetch event with blob body');

async_test(function(t) {
    var scope = 'resources/simple.html?referrer';
    service_worker_unregister_and_register(t, worker, scope)
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(t.step_func(function(sw) {
          return wait_for_state(t, sw, 'activated');
        }))
      .then(t.step_func(function() { return with_iframe(scope); }))
      .then(t.step_func(function(frame) {
          assert_equals(
            frame.contentDocument.body.textContent,
            'Referrer: ' + document.location.href,
            'Service Worker should respond to fetch with the referrer URL');
          return service_worker_unregister_and_done(t, scope);
        }))
      .catch(unreached_rejection(t));
  }, 'Service Worker responds to fetch event with the referrer URL');

async_test(function(t) {
    var scope = 'resources/simple.html?ignore';
    service_worker_unregister_and_register(t, worker, scope)
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(t.step_func(function(sw) {
          return wait_for_state(t, sw, 'activated');
        }))
      .then(t.step_func(function() { return with_iframe(scope); }))
      .then(t.step_func(function(frame) {
          assert_equals(frame.contentDocument.body.textContent,
                        'Here\'s a simple html file.\n',
                        'Response should come from fallback to native fetch');
          return service_worker_unregister_and_done(t, scope);
        }))
      .catch(unreached_rejection(t));
  }, 'Service Worker does not respond to fetch event');

async_test(function(t) {
    var scope = 'resources/simple.html?null';
    service_worker_unregister_and_register(t, worker, scope)
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(t.step_func(function(sw) {
          return wait_for_state(t, sw, 'activated');
        }))
      .then(t.step_func(function() { return with_iframe(scope); }))
      .then(t.step_func(function(frame) {
          assert_equals(frame.contentDocument.body.textContent,
                        '',
                        'Response should be the empty string');
          return service_worker_unregister_and_done(t, scope);
        }))
      .catch(unreached_rejection(t));
  }, 'Service Worker responds to fetch event with null response body');

async_test(function(t) {
    var scope = 'resources/simple.html?reject';
    service_worker_unregister_and_register(t, worker, scope)
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(t.step_func(function(sw) {
          return wait_for_state(t, sw, 'activated');
        }))
      .then(t.step_func(function() { return with_iframe(scope); }))
      .then(t.step_func(function(frame) {
          assert_equals(frame.contentDocument.body.textContent,
                        'Here\'s a simple html file.\n',
                        'Response should come from fallback to native fetch');
          return service_worker_unregister_and_done(t, scope);
        }))
      .catch(unreached_rejection(t));
  }, 'Service Worker rejects fetch event');

async_test(function(t) {
    var scope = 'resources/simple.html?fetch';
    service_worker_unregister_and_register(t, worker, scope)
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(t.step_func(function(sw) {
          return wait_for_state(t, sw, 'activated');
        }))
      .then(t.step_func(function() { return with_iframe(scope); }))
      .then(t.step_func(function(frame) {
          assert_equals(frame.contentDocument.body.textContent,
                        'Here\'s an other html file.\n',
                        'Response should come from fetched other file');
          return service_worker_unregister_and_done(t, scope);
        }))
      .catch(unreached_rejection(t));
  }, 'Service Worker fetches other file in fetch event');
</script>
</body>
